+++ Angular-Front-End/app-front-end/src/app/models/login.model.ts
@@ -1,0 +1,7 @@
+import { Identifiable } from './identifiable';
+import { UserModel } from './user.model';
+
+export interface LoginModel extends Identifiable{
+    user: UserModel;
+    msg: string;
+}