+++ Services/UserCatalogService/src/main/java/io/App/UserManagementService/userComponent/User.java
@@ -3,7 +3,9 @@ package io.App.UserManagementService.userComponent;
 public class User {
 
 	private int id;
-	private String name;
+	private String uName;
+	private String firstName;
+	private String lastName;
 	private String email;
 	private String password;
 
@@ -18,11 +20,22 @@ public class User {
 	 * @param name
 	 *            - name of the user
 	 */
-	public User(int id, String name, String email, String password) {
+	public User(int id, String name, String firstName, String lastName, String email, String password) {
 		this.id = id;
-		this.name = name;
+		this.uName = name;
+		this.firstName = firstName;
+		this.lastName = lastName;
 		this.email = email;
 		this.password = password;
+	}
+
+	/**
+	 * This contructor is used to remove users from the system
+	 * @param getuName
+	 * @param password2
+	 */
+	public User(String getuName, String password2) {
+		// TODO Auto-generated constructor stub
 	}
 
 	public int getId() {
@@ -30,7 +43,7 @@ public class User {
 	}
 
 	public String getName() {
-		return name;
+		return uName;
 	}
 
 	public String getEmail() {
@@ -39,6 +52,14 @@ public class User {
 
 	public String getPassword() {
 		return password;
+	}
+	
+	public String getFirstName() {
+		return firstName;
+	}
+	
+	public String getLastName() {
+		return lastName;
 	}
 
 }