+++ Services/UserCatalogService/src/main/java/io/App/UserManagementService/userComponent/UserCatalog.java
@@ -4,6 +4,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 import io.App.UserManagementService.databaseConnection.UserDatabaseConnection;
 import io.App.UserManagementService.dto.UserListWrapper;
+import io.App.UserManagementService.exceptions.InternalAppException;
 import io.App.UserManagementService.exceptions.UserAlreadyExistsException;
 import io.App.UserManagementService.exceptions.UserDoesNotExistException;
 
@@ -18,6 +19,7 @@ public class UserCatalog {
 
 	/**
 	 * Get all users registered
+	 * 
 	 * @return
 	 */
 	public UserListWrapper getUsers() {
@@ -26,32 +28,49 @@ public class UserCatalog {
 
 	/**
 	 * Add a user to the system
+	 * 
 	 * @param user - user to add
 	 * @throws UserAlreadyExistsException - in case the user already exists
+	 * @throws InternalAppException - in case there is an internal error with
+	 * the Application
 	 */
-	public void addUser(User user) throws UserAlreadyExistsException {
-		if(uDC.getUserByName(user.getName()) != null) {
-			throw new UserAlreadyExistsException();
-		}
+	public void addUser(User user)
+			throws UserAlreadyExistsException, InternalAppException {
 		uDC.addUser(user);
 	}
 
 	/**
 	 * Remove a user from the system
+	 * 
 	 * @param user - user to remove
-	 * @throws UserDoesNotExistException
+	 * @throws UserDoesNotExistException - in case the user doesn't exist
+	 * @throws InternalAppException - in case there is an internal error with
+	 * the Application
 	 */
-	public void removeUser(User user) throws UserDoesNotExistException {
-		if(uDC.getUserByName(user.getName()) == null) {
-			throw new UserDoesNotExistException();
-		}
+	public void removeUser(User user) throws InternalAppException {
 		uDC.removeUser(user);
 	}
 
-	public void getUserInfo(int uID) {
-		// TODO
+	/**
+	 * Gets a user with a given id
+	 * 
+	 * @param uID - id of the user to get
+	 * @return - the user with the given id
+	 */
+	public User getUserInfo(int uID) {
+		return null;
 	}
-	
-	
+
+	/**
+	 * Gets a user with a given name
+	 * 
+	 * @param uName - the name of the user to get
+	 * @return - the user with the given name
+	 * @throws InternalAppException - in case there is an internal error with
+	 * the Application
+	 */
+	public User getUserByName(String uName) throws InternalAppException {
+		return uDC.getUserByName(uName);
+	}
 
 }