+++ changelog.md
@@ -3,6 +3,33 @@ All changes made to this project will be documented in this file
 
 ## [Unreleased]
 
+## [0.1.2] - 2020-07-19
+### Changed
+- Login Page was finished. Started testing with login Backend
+- Login service is now Authentication Service
+- UserCatalog was changed and errors are now based on exception throwing
+- Changed User Model. Now has fields firstName and lastName
+- Changed UserDTO to incorporate changes made to the user model
+- Changed Database to incorporate changes made to the user model
+
+
+### Added
+- Method for user login
+- Start of Metod for user update
+- Added new class to receive data from Login page
+- "_helpers" folder with fake backend for testing
+- Added classes for interception of errors and requests to backend
+- New component events for all upcoming user events
+
+-------------------------------------------------------------------------
+
+## [0.1.2] - 2020-07-17
+### Added
+- Certificates for SSL connection establishment
+- New startcommand to test with ssl. Still not working
+
+-------------------------------------------------------------------------
+
 ## [0.1.2] - 2020-07-15
 ### Changed
 - Login Page was changed. Now trying to do login with simple mysql database
@@ -10,6 +37,7 @@ All changes made to this project will be documented in this file
 - Continued working on User Component. Signup still not working
 - Finished working on database. Everything now working
 
+-------------------------------------------------------------------------
 
 ## [0.1.2] - 2020-07-14
 ### Changed
@@ -28,6 +56,7 @@ All changes made to this project will be documented in this file
 ### Changed
 - Completed Community Service. Everything is now working (tested with Postman)
 
+-------------------------------------------------------------------------
 
 ## [0.1.2] - 2020-07-08
 ### Added
@@ -41,6 +70,7 @@ maintainers of the system have and ADMIN_ROLE, which is the role of the admin
 - Created constants for all SQL Queries
 - Improve Project Structure
 
+-------------------------------------------------------------------------
 
 ## [0.1.1] - 2020-07-07
 ### Changed